timetracker/requirements.txt
flask
pytz
werkzeug
dotenv
fpdf2
